% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportnormalitytests.R
\docType{package}
\name{reportnormalitytests}
\alias{reportnormalitytests}
\title{reportnormalitytests Package}
\description{
Automatic report of findings from various normality tests
Description: This function automatically creates reports of different normality tests. These tests are: ols_test_normality, shapiro.test, ad.tests, cvm.test, lillie.test, pearson.test, sf.test and ks.test.
}
\seealso{
Relevant functions of this package:
- {report_normality()}
- {report_ols()}
}
\author{
Anna Huelemeier \email{anna@huelemeier.de}
}
